refactor: inquiry package
