--- src/main/java/org/hoongoin/interviewbank/inquiry/controller/InquiryResponse.java
@@ -1,4 +1,0 @@
-package org.hoongoin.interviewbank.inquiry.controller;
-
-public class InquiryResponse {
-}
