+++ src/main/java/org/hoongoin/interviewbank/inquiry/infrastructure/InquiryRepository.java
@@ -1,5 +1,6 @@
 package org.hoongoin.interviewbank.inquiry.infrastructure;
 
+import org.hoongoin.interviewbank.inquiry.infrastructure.entity.InquiryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
